requirements.txt
appnope==0.1.0
attrs==19.3.0
audiolazy==0.6
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
idna==2.8
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
phue==1.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
pyzmq==18.1.1
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
Send2Trash==1.5.0
six==1.14.0
spotipy==2.10.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
